perf(zinit): postpones just some plugins and snippets initializations
This keeps all but one snippet without postponing behavior. Without this
change, the p10k behavior of transient prompt is unavailable.

Turns out postponing the initialization of the other plugins and
snippetr for after the prompt initialization does not impact on this, so
they are safe. This issue occur only for keybindings snippers. The
plugins not postponed also dont work when postponed.

https://github.com/romkatv/powerlevel10k?tab=readme-ov-file#transient-prompt